All Rive is replaced with Rive Native now
