+++ packages/podcast_core/lib/widgets/filter_episodes_popup.dart
@@ -4,7 +4,6 @@ import 'package:podcast_core/data/episodes_filter_state.dart';
 import 'package:podcast_core/extensions/text_style_extensions.dart';
 import 'package:podcast_core/providers/episodes_filter_provider.dart';
 import 'package:podcast_core/widgets/rive/new_podcast_animation.dart';
-import 'package:podcast_core/widgets/rive/podcast_animation.dart';
 
 class FilterEpisodesPopup extends ConsumerWidget {
   const FilterEpisodesPopup({super.key});
