+++ packages/podcast_core/lib/widgets/small_media_player/download_animation.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:podcast_core/data/episode.model.dart';
 import 'package:podcast_core/providers/episode_loader_provider.dart';
-import 'package:podcast_core/widgets/rive/podcast_animation.dart';
+import 'package:podcast_core/widgets/rive/new_podcast_animation.dart';
 
 class DownloadAnimation extends HookConsumerWidget {
   const DownloadAnimation({super.key, required this.episode});
@@ -17,7 +17,7 @@ class DownloadAnimation extends HookConsumerWidget {
       ).select((e) => e.value?.currentDownloadProgress),
     );
 
-    return PodcastAnimation(
+    return NewPodcastAnimation(
       artboard: PodcastAnimationArtboard.download,
       params: {'Progress': progress},
     );
